CRI: Let OCI runtime create the netns
Currently we create the netns via CLONE and ask the pod sandbox to join
it. This is problematic with the upcoming user-namespaces support in
Kubernetes, as the netns must be owned by the userns (so it can, if
proper capabilities are given, govern it). This is not possible today,
as we are creating the netns before the userns is created by the OCI
runtime.

This patch just moves the CNI network setup to after the pod sandbox is
created and instead of creating the netns ourselves, we let the OCI
runtime create the netns. The OCI runtime creates the netns properly,
and is owned by the userns if one is used.

As a side-effect of these changes, the netns.newNS() linux function is
simplified: all the games with the go runtime and CLONE are removed.
Those were only needed when we CLONE a new nets. We don't do that
anymore, we just create a persistent mount for the CNI to work with.

Another side effect is that the runtimeStart variable/metric in
pkg/cri/server/sandbox_run.go before this patch didn't count the time of
doing the CNI setup. After this patch, as the CNI setup is done after
the pod sandbox is started (we need to start it for it to create the
netns), now the runtimeStart includes the time of the CNI setup. This
might be worth documenting in the changelog for this release.

TODO rata: ask in the PR if this is fine or we should move some of the
startSandboxExitMonitor() or similar before the CNI network setup.

TODO rata: ask in the PR if we can/should do something else for windows.
Like make the RunPodSandbox() thing do:
 1. Create netns // no-op on platforms that allow us to join the netns later
 2. Low lever container runtime: create pod sandbox
 3. Create a new container that joins that network (created in 1 or 2,
    depends the platform. For 2 it uses the pid, for case 1, the
    function to create a netns returns a reference)
 4. Setup the pod network (CNI)
 5. Do the rest.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>