WIP: Add a capability, so snapshotters can opt-in to remapping
This allows a snapshotter to report if it can handle the user remapping
labels; therefore allowing the snapshotter to either be configured to do
so, or detect that a new enough kernel is available as specific to
the snapshotter.

If the snapshotter does not support the labels, the core does the
remapping itself, currently the implementation of this is a bit strange
as it has to expand the snapshotter's info and labels itself to check
this.

Also needs tests, clearly.